+++ runtime/benchmark_balances_force_free_test.go
@@ -6,8 +6,6 @@ import (
 	"testing"
 
 	gossamertypes "github.com/ChainSafe/gossamer/dot/types"
-	"github.com/ChainSafe/gossamer/lib/common"
-	"github.com/ChainSafe/gossamer/lib/runtime"
 	"github.com/ChainSafe/gossamer/pkg/scale"
 	sc "github.com/LimeChain/goscale"
 	"github.com/LimeChain/gosemble/primitives/benchmarking"
@@ -32,7 +30,7 @@ func benchmarkBalancesForceFree(b *testing.B) {
 	assert.NoError(b, err)
 
 	// Create the call
-	call, err := ctypes.NewCall(metadata, "Balances.force_free", alice, ctypes.NewU128(*big.NewInt(600000000)))
+	call, err := ctypes.NewCall(metadata, "Balances.force_free", alice, ctypes.NewU128(*big.NewInt(2 * existentialAmount)))
 	assert.NoError(b, err)
 
 	// Create the extrinsic
@@ -50,9 +48,6 @@ func benchmarkBalancesForceFree(b *testing.B) {
 	}
 
 	// Setup the state
-	// This setup differs from the one in the Substrate, and the
-	// total (free + reserved) is less than the existential deposit,
-	// resulting in additional read/write for "Balances TotalIssuance" key
 	pubKey := signature.TestKeyringPairAlice.PublicKey
 	accountInfo := gossamertypes.AccountInfo{
 		Nonce:       0,
@@ -60,8 +55,8 @@ func benchmarkBalancesForceFree(b *testing.B) {
 		Producers:   0,
 		Sufficients: 0,
 		Data: gossamertypes.AccountData{
-			Free:       scale.MustNewUint128(big.NewInt(500000000)),
-			Reserved:   scale.MustNewUint128(big.NewInt(500000000)),
+			Free:       scale.MustNewUint128(big.NewInt(existentialAmount)),
+			Reserved:   scale.MustNewUint128(big.NewInt(existentialAmount)),
 			MiscFrozen: scale.MustNewUint128(big.NewInt(0)),
 			FreeFrozen: scale.MustNewUint128(big.NewInt(0)),
 		},
@@ -69,8 +64,8 @@ func benchmarkBalancesForceFree(b *testing.B) {
 	setAccountInfo(b, storage, pubKey, accountInfo)
 
 	info := getAccountInfo(b, storage, pubKey)
-	assert.Equal(b, scale.MustNewUint128(big.NewInt(500000000)), info.Data.Reserved)
-	assert.Equal(b, scale.MustNewUint128(big.NewInt(500000000)), info.Data.Free)
+	assert.Equal(b, scale.MustNewUint128(big.NewInt(existentialAmount)), info.Data.Reserved)
+	assert.Equal(b, scale.MustNewUint128(big.NewInt(existentialAmount)), info.Data.Free)
 
 	(*storage).DbWhitelistKey(string(append(keySystemHash, keyBlockNumberHash...)))    // 1 read/write
 	(*storage).DbWhitelistKey(string(append(keySystemHash, keyExecutionPhaseHash...))) // 1 read
@@ -84,7 +79,7 @@ func benchmarkBalancesForceFree(b *testing.B) {
 	// Validate the result/state
 	info = getAccountInfo(b, storage, pubKey)
 	assert.Equal(b, scale.MustNewUint128(big.NewInt(0)), info.Data.Reserved)
-	assert.Equal(b, scale.MustNewUint128(big.NewInt(2*500000000)), info.Data.Free)
+	assert.Equal(b, scale.MustNewUint128(big.NewInt(2*existentialAmount)), info.Data.Free)
 
 	benchmarkResult, err := benchmarking.DecodeBenchmarkResult(bytes.NewBuffer(res))
 	assert.NoError(b, err)
@@ -100,42 +95,4 @@ func benchmarkBalancesForceFree(b *testing.B) {
 	b.Cleanup(func() {
 		rt.Stop()
 	})
-}
-
-func setAccountInfo(b *testing.B, storage *runtime.Storage, account []byte, info gossamertypes.AccountInfo) {
-	bytesStorage, err := scale.Marshal(info)
-	assert.NoError(b, err)
-
-	err = (*storage).Put(accountStorageKey(account), bytesStorage)
-	assert.NoError(b, err)
-}
-
-func getAccountInfo(b *testing.B, storage *runtime.Storage, account []byte) *gossamertypes.AccountInfo {
-	accountInfo := gossamertypes.AccountInfo{
-		Nonce:       0,
-		Consumers:   0,
-		Producers:   0,
-		Sufficients: 0,
-		Data: gossamertypes.AccountData{
-			Free:       scale.MustNewUint128(big.NewInt(0)),
-			Reserved:   scale.MustNewUint128(big.NewInt(0)),
-			MiscFrozen: scale.MustNewUint128(big.NewInt(0)),
-			FreeFrozen: scale.MustNewUint128(big.NewInt(0)),
-		},
-	}
-
-	bytesStorage := (*storage).Get(accountStorageKey(account))
-
-	err := scale.Unmarshal(bytesStorage, &accountInfo)
-	assert.NoError(b, err)
-
-	return &accountInfo
-}
-
-func accountStorageKey(account []byte) []byte {
-	aliceHash, _ := common.Blake2b128(account)
-	keyStorageAccount := append(keySystemHash, keyAccountHash...)
-	keyStorageAccount = append(keyStorageAccount, aliceHash...)
-	keyStorageAccount = append(keyStorageAccount, account...)
-	return keyStorageAccount
 }
